Uninstaller now deletes Uninstall registry keys
