+++ Uninstaller/MainWindow.resx
