+++ Uninstaller/Program.cs
@@ -32,7 +32,7 @@ namespace Uninstaller
             {
                 Application.EnableVisualStyles();
                 Application.SetCompatibleTextRenderingDefault(false);
-                Application.Run(new Form1());
+                Application.Run(new MainWindow());
 
             }
